+++ kpi-tracker/resources/schema.sql
@@ -36,7 +36,8 @@ CREATE TABLE IF NOT EXISTS paydays (
   block INT8 PRIMARY KEY REFERENCES blocks(id) ON DELETE RESTRICT ON UPDATE RESTRICT,
   total_stake INT8 NOT NULL,
   num_bakers INT8 NOT NULL,
-  num_finalizers INT8 NOT NULL
+  num_finalizers INT8 NOT NULL,
+  num_delegators INT8 NOT NULL
 );
 
 -- All accounts created.