+++ kpi-tracker/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for the kpi-tracker service
 
+## Unreleased changes
+
+- Split transaction graphs into separate types of transactions.
+- Add graph for accounts with CCD transfers.
+
 ## 1.1.0
 
 - Support protocol version 6.