+++ kpi-tracker/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Split transaction graphs into separate types of transactions.
 - Add graph for accounts with CCD transfers.
 - Add graph for number of active bakers.
+- Add graph for number of active finalizers.
 
 ## 1.1.0