+++ kpi-tracker/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 - Split transaction graphs into separate types of transactions.
 - Add graph for accounts with CCD transfers.
-- Add graph for number of active bakers.
 - Add graph for number of active finalizers.
+- Add graph showing active staking accounts, divided into bakers and delegators.
 
 ## 1.1.0