+++ kpi-tracker/CHANGELOG.md
@@ -4,7 +4,12 @@
 
 - Split transaction graphs into separate types of transactions.
 - Add graph for accounts with CCD transfers.
-- Add graph for number of active bakers.
+- Add graph for number of active finalizers.
+- Add graph showing active staking accounts, divided into bakers and delegators.
+- Add graph showing status of active bakers.
+- Add graph of delegation recipients.
+- Add graph for minted CCDs.
+- Add graphs for baker, finalizer, and foundation rewards.
 
 ## 1.1.0